requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
weasyprint==62.3
# WeasyPrint 62.3 calls the old pydyf Stream.transform() API; pydyf>=0.11
# removed it. Pin explicitly because weasyprint's own requirements leave
# the upper bound open, so pip resolves to an incompatible release.
pydyf==0.10.0
bcrypt==4.1.3
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.2.2
pypdf==5.1.0
